Update the help text and improve tool outputs
